Prepared the Character for Gameplay Ability System
Trough this arc of work, I started the work on the Gameplay Ability System for Main Character. Made a set up for Ability Activation Policy, made a Start Up Data Asset for both Main Character and Enemy and Defined it for Main Character. Set up a Weapon Class for Main Character and granted the ability to Main Character to Spawn it at the start of the game.