Unequip Ability Progress
In this arc I've implemented custom Weapon Abilities and Weapon Input Mapping Context in purpose to have a possibility to override default Abilities and Input corresponding to specific Weapon. Now if we start the game and press 1, our Character will run Equip Animation and Equip the Weapon and if we press 1 again we will see Unequip Animation. Now main task to Unequip the Weapon and work on Weapon Abilities and Input.