Successful Unequipping the Weapon
Now, after we already have a Weapon Equipped, if we press 1 again we are Unequipping our Weapon, cleaning Weapon Input Mapping Context and replacing it with default and also cleaning Weapon Abilities and changing it back to Default Abilities. These steps allows us to Equip and Unequip our weapon and get a different Abilities based on Equip status. Next task is to Develop an Attack Abilities.