Enemy Character
During this arc I've started the work on the Enemy Character. I made a basic setup for Base Enemy, as well as created an instance of it and called it Gruntling Guardian. It would be first and weakest type of Enemy. During next arc of work main task is to finish setup of an Enemy and Implement Attribute System in purpose to start implementing Damage System.