Damage Effect Calculation
During this arc I've implemented the Calculation of the Damage Effect. The Calculation is done for Light and Heavy Attack. Damage depends on the current difficulty. Damage also includes current characteristics of a Target, such as Defense Power. During next arc main task is to modify Health corresponding to the final damage and implement Hit Effect.