Final Damage and Hit Effects
During this arc I've finally added calculation of a Damage based on Final Damage and corresponding change of a Health Attribute on Hit. Now whenever we Hit our Enemy, it's health attribute is modified in response to Final Damage. Also started the work on the Hit Effects. For now I have a Hit Effect for Enemy Character. When we Hit our Enemy it changes it's color to reddish for a small amount of time. Also I've added a small Pause Effect on Hit, which improves overall feeling of the Hit.