Started the work on Widgets
During this arc I've started the work on the Widgets for both Hero and Enemy characters. I've implemented broadcasting of the characteristics changes in both Hero and Enemy classes, also implemented listening for the changes in the broadcast for Widget Class. Also added a test widget to ensure, that both Hero and Enemy characters broadcasting correctly and Widget Class listening for the changes. During next arc main task is to implement actual Widgets, that would be used in future.