Work on Enemy AI
During this arc I've started the work on the Enemy UI Component. I've made a basic setup for Crowd Following Component, AI Perception. Made a setup for Generic Team ID in purpose for AI to differ actors on the level. And created a basic Behavior Tree with basic setup. Next task is to dig further into Enemy AI and setup more advanced features.