Enemy Death Ability and Ability System Fix
During this arc I've implemented the Hero Death Ability with corresponding montages to play. Also after Hero Death we turn off collision and input ability for player and turning on the cursor. Also I fond a fix for Ability System for 5.5 and higher. Previously without fix our Block Ability with Tags didn't work correctly and it didn't actually blocked abilities and because of that we had a lot of bugs. But after overriding DoesAbilitySatisfyTagRequirements() function, everything works like it should. Next task is to implement new type of an Enemy.