Rage implementation Part 2
During this arc I've finished the implementation of Rage Ability. Now after our Rage Ability is activated our Attacks are getting buffed and now shooting special Hero Projectile, that pierces all Enemy Characters on the way and damaging them. Also implemented the possibility to restore Rage for Hero Character. Now when we Attack enemy with either Light or Heavy Attacks we are restoring some part of our Rage. For Light Attack this amount of restored Rage is smaller, than for Heavy Attack. Next task is to start the implementation of Hero Special Abilities.