Hero Special Ability and Custom Cooldown
During this arc I've implemented the Hero Special Ability, which can be activated with pressing Alt + LMB. Also created corresponding Anim Montages, Sound Notifies and Hit Notifies. Hero Special Abilities have cooldown, so I had to create a Custom Latent Action, which in out case is Count Down Action. This Count Down Action is Customizable and can be used in many cases. In my game it is used to track the cooldown of Special Ability and to show it to the player. Also I managed the case when Player Character hide his weapon and after equipping it again count down wasn't displayed. Now after character unequipping and equipping the weapon ability cooldown is still going on and is displayed correctly. Next task is to implement the Heavy Special Ability.