Added Playable map for Main Project and Maid a Game Mode Setup
During this arc I've created the playable Map for main project. In this map I've created the Arena, where both main character and enemies spawns. Checked the arena for bugs and ensured that there is no way for character to go Outside the Arena. Setup the Enemies Spawn Points and Navigation Mesh.