Small Items change
During this arc of work I've changed a little bit the Item Description it self. Also modified Potions Item Descriptions to make them look good and correspond to the In-Game lore. Also changed the Cloak sizing to make it more suitable for the Character Model. Also added the new Mask Equippable and tested it's functionality in the game.