Clean Up
During this small arc I've cleaned up the .h and .cpp files, deleted all unnecessary Logs and On Screen Debug Messages. Also made some small tweaks inside the Character Input in purpose for it to correspond to the In-Game UI hints.

Next task is to implement - 1. In-Game Controls Hint Widget. 2. Add the Sensitivity Setting 3. Add the Volume Setting