Item Pick Up with an Overlap Event
During this arc I've added another way of picking up an Item. Previously it was possible only through the Line Tracing, which was uncomfortable for some of the Testers. Now I've added the possibility to pick up an Item by Overlapping with it. So now player can choose the way how to pick up an Item.